+++ modules/main/src/main/java/com/nisovin/shopkeepers/shopkeeper/player/book/BookPlayerShopEditorHandler.java
@@ -14,6 +14,7 @@ import org.bukkit.inventory.meta.BookMeta;
 import com.nisovin.shopkeepers.api.shopkeeper.offers.BookOffer;
 import com.nisovin.shopkeepers.api.util.UnmodifiableItemStack;
 import com.nisovin.shopkeepers.config.Settings;
+import com.nisovin.shopkeepers.config.Settings.DerivedSettings;
 import com.nisovin.shopkeepers.shopkeeper.TradingRecipeDraft;
 import com.nisovin.shopkeepers.shopkeeper.player.PlayerShopEditorHandler;
 import com.nisovin.shopkeepers.ui.editor.DefaultTradingRecipesAdapter;
@@ -113,6 +114,16 @@ public class BookPlayerShopEditorHandler extends PlayerShopEditorHandler {
 	}
 
 	@Override
+	protected TradingRecipeDraft getEmptyTrade() {
+		return DerivedSettings.bookEmptyTrade;
+	}
+
+	@Override
+	protected TradingRecipeDraft getEmptyTradeSlotItems() {
+		return DerivedSettings.bookEmptyTradeSlotItems;
+	}
+
+	@Override
 	protected void handleTradesClick(EditorSession editorSession, InventoryClickEvent event) {
 		assert this.isTradesArea(event.getRawSlot());
 		Inventory inventory = editorSession.getInventory();
@@ -121,12 +132,16 @@ public class BookPlayerShopEditorHandler extends PlayerShopEditorHandler {
 			// Change the low cost, if this column contains a trade:
 			ItemStack resultItem = this.getTradeResultItem(inventory, this.getTradeColumn(rawSlot));
 			if (resultItem == null) return;
-			this.handleUpdateTradeCostItemOnClick(event, Settings.createCurrencyItem(1), Settings.createZeroCurrencyItem());
+
+			UnmodifiableItemStack emptySlotItem = this.getEmptyTradeSlotItems().getItem1();
+			this.updateTradeCostItemOnClick(event, Settings.createCurrencyItem(1), emptySlotItem);
 		} else if (this.isItem2Row(rawSlot)) {
 			// Change the high cost, if this column contains a trade:
 			ItemStack resultItem = this.getTradeResultItem(inventory, this.getTradeColumn(rawSlot));
 			if (resultItem == null) return;
-			this.handleUpdateTradeCostItemOnClick(event, Settings.createHighCurrencyItem(1), Settings.createHighZeroCurrencyItem());
+
+			UnmodifiableItemStack emptySlotItem = this.getEmptyTradeSlotItems().getItem2();
+			this.updateTradeCostItemOnClick(event, Settings.createHighCurrencyItem(1), emptySlotItem);
 		}
 		// Result item row: Result items (books) are not modifiable.
 	}